requirements.txt
aiohttp
aiosqlite
asyncio
discord.py
pyyaml
DateTime
pytz